moves calculation to the filter util, renames state
