+++ src/Component/RuleTable/RuleTable.spec.tsx
@@ -60,32 +60,4 @@ describe('RuleTable', () => {
     });
   });
 
-  describe('calculateCountAndDuplicates', () => {
-    it('returns the right number of duplicates', () => {
-      let dummyData = TestUtil.getComplexGsDummyData();
-      let filter2 = TestUtil.getDummyGsFilter();
-
-      dummyData.exampleFeatures.features[0].properties.state = 'germany';
-      dummyData.exampleFeatures.features[0].properties.population = 150000;
-      dummyData.exampleFeatures.features[0].properties.name = 'NotSchalke';
-
-      const rules: Rule[] = [{
-        name: 'rule1',
-        symbolizers: [],
-        filter: []
-      }, {
-        name: 'rule2',
-        symbolizers: [],
-        filter: filter2
-      }];
-
-      const result = RuleTable.calculateCountAndDuplicates(rules, dummyData);
-
-      expect(result.counts[0]).toBeCloseTo(10);
-      expect(result.counts[1]).toBeCloseTo(1);
-      expect(result.duplicates[0]).toBeCloseTo(1);
-      expect(result.duplicates[1]).toBeCloseTo(1);
-    });
-  });
-
 });
